# app/routers/receipts.py
from uuid import UUID

from fastapi import APIRouter, Depends, File, Header, HTTPException, Query, Response, UploadFile, status
from pymongo.database import Database

from app import schemas, services
from app.dependencies import get_actor_user_id, get_db, get_s3

router = APIRouter(tags=["Receipts"])


@router.post(
    "/api/events/{id}/receipts",
    response_model=schemas.Receipt,
    status_code=status.HTTP_201_CREATED,
)
def create_receipt(
    id: UUID,
    payload: schemas.CreateReceiptRequest,
    idempotency_key: str = Header(min_length=1, alias="Idempotency-Key"),
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.create_receipt(
        db, str(id), payload, current_user_id, idempotency_key=idempotency_key
    )


@router.get("/api/events/{id}/receipts", response_model=schemas.ReceiptPage)
def list_receipts_by_event(
    id: UUID,
    limit: int = Query(default=50, ge=1, le=100),
    offset: int = Query(default=0, ge=0),
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.list_receipts_by_event(
        db, str(id), current_user_id, limit=limit, offset=offset
    )


@router.post(
    "/api/events/{id}/receipt-drafts/ai",
    response_model=schemas.ReceiptAIDraftResponse,
    status_code=status.HTTP_201_CREATED,
)
def create_receipt_ai_draft(
    id: UUID,
    payload: schemas.ReceiptAIDraftRequest,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.create_receipt_ai_draft(db, str(id), payload, current_user_id)


@router.get("/api/receipts/{id}", response_model=schemas.Receipt)
def get_receipt(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.get_receipt(db, str(id), current_user_id)


@router.patch("/api/receipts/{id}", response_model=schemas.Receipt)
def update_receipt(
    id: UUID,
    payload: schemas.UpdateReceiptRequest,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.update_receipt(db, str(id), payload, current_user_id)


@router.post("/api/receipts/{id}/confirm", response_model=schemas.Receipt)
def confirm_receipt(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.confirm_receipt(db, str(id), current_user_id)


@router.get(
    "/api/receipts/{id}/confirm/confirmation-summary",
    response_model=schemas.ConfirmationSummary,
)
def get_receipt_confirm_confirmation_summary(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.get_receipt_confirm_confirmation_summary(db, str(id), current_user_id)


@router.post("/api/receipts/{id}/validate", response_model=schemas.Receipt)
def validate_receipt(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.validate_receipt(db, str(id), current_user_id)


@router.get("/api/receipts/{id}/share-reviews", response_model=schemas.ReceiptShareReviewPage)
def list_receipt_share_reviews(
    id: UUID,
    limit: int = Query(default=50, ge=1, le=100),
    offset: int = Query(default=0, ge=0),
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.list_receipt_share_reviews(
        db, str(id), current_user_id, limit=limit, offset=offset
    )


@router.post("/api/receipts/{id}/share-reviews/me/accept", response_model=schemas.ReceiptShareReview)
def accept_receipt_share_review(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.accept_receipt_share_review(db, str(id), current_user_id)


@router.post("/api/receipts/{id}/share-reviews/me/dispute", response_model=schemas.ReceiptShareReview)
def dispute_receipt_share_review(
    id: UUID,
    payload: schemas.ReceiptShareReviewDispute,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.dispute_receipt_share_review(db, str(id), payload, current_user_id)


@router.post("/api/receipts/{id}/void", response_model=schemas.Receipt)
def void_receipt(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.void_receipt(db, str(id), current_user_id)


@router.get(
    "/api/receipts/{id}/void/confirmation-summary",
    response_model=schemas.ConfirmationSummary,
)
def get_receipt_void_confirmation_summary(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.get_receipt_void_confirmation_summary(db, str(id), current_user_id)


@router.post("/api/receipts/{id}/corrections", response_model=schemas.Receipt)
def create_receipt_correction(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.create_receipt_correction(db, str(id), current_user_id)


@router.post("/api/receipts/{id}/allocation-session", response_model=schemas.AllocationSessionState)
def start_allocation_session(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.start_allocation_session(db, str(id), current_user_id)


@router.get(
    "/api/allocation-sessions/{id}",
    response_model=schemas.AllocationSessionState,
)
def get_allocation_session(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.get_allocation_session(db, str(id), current_user_id)


@router.post("/api/allocation-sessions/{id}/claims", response_model=schemas.ReceiptItemClaim)
def claim_receipt_item(
    id: UUID,
    payload: schemas.ReceiptItemClaimRequest,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.claim_receipt_item(db, str(id), payload, current_user_id)


@router.delete("/api/allocation-sessions/{id}/claims", status_code=status.HTTP_204_NO_CONTENT)
def unclaim_receipt_item(
    id: UUID,
    payload: schemas.ReceiptItemClaimRequest,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> Response:
    services.unclaim_receipt_item(db, str(id), payload, current_user_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post(
    "/api/allocation-sessions/{id}/ready",
    response_model=schemas.AllocationSessionState,
)
def mark_allocation_session_ready(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.mark_allocation_session_ready(db, str(id), current_user_id)


@router.post("/api/allocation-sessions/{id}/finalize", response_model=schemas.Receipt)
def finalize_allocation_session(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict:
    return services.finalize_allocation_session(db, str(id), current_user_id)


@router.delete("/api/receipts/{id}", status_code=status.HTTP_204_NO_CONTENT)
def delete_receipt(
    id: UUID,
    db: Database = Depends(get_db),
    current_user_id: str = Depends(get_actor_user_id),
) -> Response:
    services.delete_receipt(db, str(id), current_user_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post(
    "/api/receipts/{id}/image",
    response_model=schemas.ReceiptImageUploadResponse,
    status_code=status.HTTP_201_CREATED,
)
async def upload_receipt_image(
    id: UUID,
    db: Database = Depends(get_db),
    s3=Depends(get_s3),
    current_user_id: str = Depends(get_actor_user_id),
    file: UploadFile | None = File(
        None,
        description="JPEG image (.jpg or .jpeg); use this field or `image`.",
    ),
    image: UploadFile | None = File(
        None,
        description="Same as `file` (alternate form field name some clients use).",
    ),
) -> dict[str, str]:
    upload = file or image
    if upload is None:
        raise HTTPException(
            status_code=422,
            detail="Send the JPEG as multipart form-data with field name 'file' or 'image'.",
        )
    body = await upload.read()
    return services.upload_receipt_image(
        db, s3, str(id), body, upload.content_type, current_user_id
    )


@router.delete("/api/receipts/{id}/image", status_code=status.HTTP_204_NO_CONTENT)
def delete_receipt_image(
    id: UUID,
    db: Database = Depends(get_db),
    s3=Depends(get_s3),
    current_user_id: str = Depends(get_actor_user_id),
) -> Response:
    services.delete_receipt_image(db, s3, str(id), current_user_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.get(
    "/api/receipts/{id}/image/presigned-url",
    response_model=schemas.ReceiptImagePresignedUrlResponse,
)
def get_receipt_image_presigned_url(
    id: UUID,
    db: Database = Depends(get_db),
    s3=Depends(get_s3),
    current_user_id: str = Depends(get_actor_user_id),
) -> dict[str, str]:
    return services.get_receipt_image_presigned_url(db, s3, str(id), current_user_id)